refactor: move out code from app.py
